refactor: 로또 번호 갯수를 LottoNumbers가 담당하도록 수정
