+++ src/main/java/lotto/domain/Lotto.java
@@ -1,8 +1,6 @@
 package lotto.domain;
 
 public class Lotto {
-	public static final int SIZE = 6;
-
 	private final LottoNumbers lottoNumbers;
 
 	public Lotto(final LottoNumbers lottoNumbers) {
@@ -12,18 +10,11 @@ public class Lotto {
 
 	private void validate(LottoNumbers lottoNumbers) {
 		validateNull(lottoNumbers);
-		validateSizeMismatch(lottoNumbers);
 	}
 
 	private void validateNull(LottoNumbers lottoNumbers) {
 		if (lottoNumbers == null) {
 			throw new IllegalArgumentException("null은 들어올 수 없습니다!");
-		}
-	}
-
-	private void validateSizeMismatch(LottoNumbers lottoNumbers) {
-		if (lottoNumbers.size() != SIZE) {
-			throw new IllegalArgumentException("로또 번호는 " + SIZE + "개여야 합니다!");
 		}
 	}
