+++ src/main/java/lotto/domain/LottoMachine.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class LottoMachine {
-	private final LottoNumbers allLottoNumbers;
+	private final List<LottoNumber> lottoBalls;
 
 	private LottoMachine() {
 		List<LottoNumber> initialSetting = new ArrayList<>();
@@ -14,7 +14,7 @@ public class LottoMachine {
 		for (int i = LottoNumber.MINIMUM; i <= LottoNumber.MAXIMUM; i++) {
 			initialSetting.add(new LottoNumber(i));
 		}
-		this.allLottoNumbers = new LottoNumbers(initialSetting);
+		this.lottoBalls = initialSetting;
 	}
 
 	public static LottoMachine getInstance() {
@@ -31,9 +31,9 @@ public class LottoMachine {
 	}
 
 	private List<LottoNumber> pickRandomNumbers() {
-		Collections.shuffle(allLottoNumbers.getLottoNumbers());
-		return allLottoNumbers.getLottoNumbers().stream()
-			.limit(Lotto.SIZE)
+		Collections.shuffle(lottoBalls);
+		return lottoBalls.stream()
+			.limit(LottoNumbers.SIZE)
 			.sorted()
 			.collect(Collectors.toList());
 	}
